# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企业项目知识库平台",
    "description": "Phase 1 + Phase 2 完整 API:\n\n**Phase 1:**\n- 多模态文档转换管道 (PDF/Word/Excel/PPT/图片/音视频/网页 → Markdown)\n- 基础 RAG 检索 (BGE-M3 三合一 + Qdrant + RRF + Reranker)\n- MCP 工具化查询 (list/outline/search/read 四工具)\n- 引用溯源 (内联引用标记 + Markdown 章节跳转 + 源文件定位)\n\n**Phase 2:**\n- Agentic RAG (LangGraph Plan-and-Execute, 最多4步, 推理透明化)\n- Skills 体系 (文档分析/跨文档对比/项目健康/报告生成)\n- SSE 流式输出 (GET /api/v1/search/stream, 感知延迟↓80%)\n- Contextual Retrieval (chunk 上下文前缀, 检索失败↓67%)",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/graph/entities/{project_id}": {
      "get": {
        "tags": [
          "GraphRAG"
        ],
        "summary": "List Entity Types",
        "description": "获取项目中所有实体类型及数量",
        "operationId": "list_entity_types_api_v1_graph_entities__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/neighbors": {
      "get": {
        "tags": [
          "GraphRAG"
        ],
        "summary": "Get Neighbors",
        "description": "查找实体的邻居（1-3 度）",
        "operationId": "get_neighbors_api_v1_graph_neighbors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Name"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 1,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/query": {
      "get": {
        "tags": [
          "GraphRAG"
        ],
        "summary": "Graph Query",
        "description": "自然语言图查询",
        "operationId": "graph_query_api_v1_graph_query_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Query"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/extract/{doc_id}": {
      "post": {
        "tags": [
          "GraphRAG"
        ],
        "summary": "Trigger Extraction",
        "description": "手动触发单个文档的实体关系抽取",
        "operationId": "trigger_extraction_api_v1_graph_extract__doc_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doc Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/project/{project_id}": {
      "delete": {
        "tags": [
          "GraphRAG"
        ],
        "summary": "Delete Project Graph",
        "description": "删除项目的全部图数据",
        "operationId": "delete_project_graph_api_v1_graph_project__project_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}